requirements.txt
chai
django
djangorestframework
google-api-python-client
MySQL-python
oauth2client
